Update-1

+++ Python_Basics/Chapter-2/01_Practice-Sets/PS1.py
@@ -1,7 +1,7 @@
 # Write a python program to add two numbers.
 
-a = int(input("Input Value Of 'A' : "))
-b = int(input("Input Value Of 'B' : "))
+a = float(input("Input Value Of 'A' : "))
+b = float(input("Input Value Of 'B' : "))
 
 sum = a + b
